Tidy up the helpers in SierraRequestsService test
